now kg are float

+++ db/migrate/20181028143409_add_kg_recycled_paper_to_collection_point.rb
@@ -1,5 +1,5 @@
 class AddKgRecycledPaperToCollectionPoint < ActiveRecord::Migration[5.2]
   def change
-    add_column :collection_points, :kg_recycled_paper, :int
+    add_column :collection_points, :kg_recycled_paper, :float
   end
 end
