+++ db/migrate/20181028143555_add_kg_recycled_glass_to_collection_point.rb
@@ -1,5 +1,5 @@
 class AddKgRecycledGlassToCollectionPoint < ActiveRecord::Migration[5.2]
   def change
-    add_column :collection_points, :kg_recycled_glass, :int
+    add_column :collection_points, :kg_recycled_glass, :float
   end
 end
